+++ modules/ticket/models/search/TicketSearch.php
@@ -9,6 +9,8 @@ use yii\data\ActiveDataProvider;
 use app\modules\ticket\models\Ticket;
 use app\components\helpers\DbHelper;
 use app\modules\sale\models\Customer;
+use yii\db\Query;
+use yii\db\Expression;
 
 /**
  * TicketSearch represents the model behind the search form about `app\modules\agenda\models\Ticket`.
@@ -44,7 +46,7 @@ class TicketSearch extends Ticket {
             [['ticket_id'], 'integer'],
             [['title', 'start_date', 'start_date', 'finish_date', 'status_id', 'customer_id', 'color_id', 'category_id', 'number', 'customer', 'document', 'assignations', 'customer_number'], 'safe', 'on' => 'wideSearch'],
             [['title', 'start_date', 'customer_id', 'color_id', 'number', 'customer_number'], 'safe', 'on' => 'activeSearch'],
-            [['search_text', 'ticket_management_qty', 'close_from_date', 'close_to_date', 'category_id', 'categories', 'customer_id', 'created_by', 'start_date_from', 'start_date_to'], 'safe'],
+            [['search_text', 'ticket_management_qty', 'close_from_date', 'close_to_date', 'category_id', 'categories', 'customer_id', 'created_by', 'start_date_from', 'start_date_to', 'status_id', 'assignations'], 'safe'],
         ];
     }
     
@@ -327,5 +329,60 @@ class TicketSearch extends Ticket {
         }
 
         return $query;
+    }
+
+    /**
+     * Devuelve registros para el reporte.
+     */
+    public function searchReport($params)
+    {
+        $this->load($params);
+
+        $query = new Query();
+        $query->select([new Expression('date_format(start_date, \'%Y-%m\') AS periodo'),new Expression('count(*) as qty')])
+            ->from('ticket');
+
+        if($this->assignations) {
+            $query->leftJoin('assignation assig', 'assig.ticket_id = ticket.ticket_id')
+                ->andFilterWhere(['assig.user_id' => $this->assignations]);
+        }
+
+        if($this->ticket_management_qty) {
+            $ticket_managements_ids = (new Query)
+                ->select('tm.ticket_id')
+                ->from('ticket_management tm')
+                ->groupBy('tm.ticket_id')
+                ->having("count(tm.ticket_id) = $this->ticket_management_qty");
+
+            $query->andFilterWhere(['in', 'ticket.ticket_id', $ticket_managements_ids]);
+        }
+
+        if($this->start_date_from){
+            $query->andWhere(['>=','start_date', $this->start_date_from]);
+        }
+
+        if($this->start_date_to) {
+            $query->andWhere(['<=', 'start_date', $this->start_date_to]);
+        }
+
+        if($this->status_id) {
+            $query->andWhere(['status_id' => $this->status_id]);
+        }
+
+        if($this->created_by) {
+            $query->andFilterWhere(['ticket.user_id' => $this->created_by]);
+        }
+
+        if($this->user_id) {
+            $query->andFilterWhere(['like', 'user.id', $this->user_id]);
+        }
+
+        if($this->category_id) {
+            $query->andFilterWhere(['category_id' => $this->category_id]);
+        }
+
+        $query->groupBy(new Expression('date_format(start_date, \'%Y-%m\')'));
+
+        return $query->all(Yii::$app->dbticket);
     }
 }